+++ app/components/CircleIconButton/tests/index.test.js
@@ -14,4 +14,17 @@ describe('<CircleIconButton />', () => {
     );
     expect(rendered).toMatchSnapshot();
   });
+
+  it('should correctly render HTML (or a component) that was inserted inside of the component using the innerElement prop', () => {
+    const testInnerElement = <div>Inner Element</div>;
+    const rendered = shallow(
+      <CircleIconButton
+        id="mockId"
+        name="config"
+        onClick={() => {}}
+        innerElement={testInnerElement}
+      />
+    );
+    expect(rendered).toMatchSnapshot();
+  });
 });