+++ Changelog.md
@@ -2,6 +2,7 @@
 * CXV1-9000 - Added Context Flag based on url + tests
 * CXV1-9001 - Toolbar styling for Login
 * CXV1-9003 - Toolbar styling for Messaging interactions (templates)
+* CXV1-9292 - Modified circle icon component so that a popup component could be inserted inside of it and retain its position relative to the button (as opposed to making dynamic positioning changes via JS if the button's position changes)
 
 ### [1.0.1] (2017-07-18)
 * CXV1-10571 - Fix critical error on contact history from click to email